+++ java-src/preflex/instrument/jdbc/DataSourceWrapper.java
@@ -18,28 +18,27 @@ import java.util.logging.Logger;
 
 import javax.sql.DataSource;
 
-import preflex.instrument.EventHandlerFactory;
 import preflex.instrument.task.CallTask1;
-import preflex.instrument.task.InstrumentingWrapper;
+import preflex.instrument.task.Wrapper;
 
 public class DataSourceWrapper<JdbcConnectionCreation, JdbcStatementCreation, SQLExecution> implements DataSource {
 
     private final DataSource ds;
     private final JdbcEventFactory<JdbcConnectionCreation, JdbcStatementCreation, SQLExecution> eventFactory;
-    private final InstrumentingWrapper<JdbcConnectionCreation> connCreationWrapper;
-    private final EventHandlerFactory<JdbcStatementCreation> stmtCreationListener;
-    private final EventHandlerFactory<SQLExecution> sqlExecutionListener;
+    private final Wrapper<JdbcConnectionCreation> connCreationWrapper;
+    private final Wrapper<JdbcStatementCreation> stmtCreationListener;
+    private final Wrapper<SQLExecution> sqlExecutionListener;
 
     public DataSourceWrapper(DataSource ds,
             final JdbcEventFactory<JdbcConnectionCreation, JdbcStatementCreation, SQLExecution> eventFactory,
-            final EventHandlerFactory<JdbcConnectionCreation> connCreationListener,
-            final EventHandlerFactory<JdbcStatementCreation> stmtCreationListener,
-            final EventHandlerFactory<SQLExecution> sqlExecutionListener) {
+            final Wrapper<JdbcConnectionCreation> connCreationWrapper,
+            final Wrapper<JdbcStatementCreation> stmtCreationWrapper,
+            final Wrapper<SQLExecution> sqlExecutionWrapper) {
         this.ds = ds;
         this.eventFactory = eventFactory;
-        this.connCreationWrapper = new InstrumentingWrapper<>(connCreationListener);
-        this.stmtCreationListener = stmtCreationListener;
-        this.sqlExecutionListener = sqlExecutionListener;
+        this.connCreationWrapper = connCreationWrapper;
+        this.stmtCreationListener = stmtCreationWrapper;
+        this.sqlExecutionListener = sqlExecutionWrapper;
     }
 
     @Override